refactor(utils): rename `get-snark-artifacts` to `snark-artifacts`
And colocate module tests in same test file
