+++ packages/utils/src/index.ts
@@ -12,5 +12,5 @@ export * from "./conversions"
 export * from "./proof-packing"
 export * from "./type-checks"
 export * from "./error-handlers"
-export * from "./get-snark-artifacts/get-snark-artifacts.node"
-export { GetSnarkArtifactUrls } from "./get-snark-artifacts/config"
+export * from "./snark-artifacts/snark-artifacts.node"
+export { GetSnarkArtifactUrls } from "./snark-artifacts/config"
