+++ packages/utils/src/snark-artifacts/snark-artifacts.browser.ts
