+++ packages/utils/src/snark-artifacts/snark-artifacts.node.ts
